requirements.txt
# Основные зависимости
aiofiles==25.1.0
aiogram==3.24.0
aiohttp==3.13.3
fastapi==0.128.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
uvicorn==0.40.0
transformers==5.0.0
torch==2.9.1
numpy==2.4.2
accelerate==1.12.0

# Зависимости для разработки
pytest==9.0.2
black==24.3.0
flake8==7.0.0
isort==5.13.2
